Add date utilities and refactor chart component
